pyproject.toml
[tool.poetry]
name = "home-work-1"
version = "0.1.0"
description = ""
authors = ["Max Kuvshinov <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
chardet = "^5.2.0"


[tool.poetry.group.dev.dependencies]
requests = "^2.31.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"

[tool.isort]
profile = "hug"
src_paths = ["isort", "test"]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cach
    | \.tox
    | \.venv
    | dist
  )/
)
'''